requirements.txt
pytest
allure-pytest
selene>=2.0.b16
smsactivate
python-dotenv